feat: Article image section resource
