+++ src/components/Link/index.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 
 import type { Article, Page, Post } from '@/payload-types'
 
-type CMSLinkType = {
+export type CMSLinkType = {
   appearance?: 'inline' | ButtonProps['variant']
   children?: React.ReactNode
   className?: string
